interfaz conectada con Base de Datos Funcional!

+++ primer_final/src/primer_final/Cuenta.java
@@ -1,12 +1,15 @@
 package primer_final;
+import java.math.BigDecimal;
+
+
 public class Cuenta{
-    private int id_cuenta;
-    private int pin_cuenta;
-    private double saldo_cuenta;
-    private int id_cliente;
+    private long id_cuenta;
+    private long pin_cuenta;
+    private BigDecimal saldo_cuenta;
+    private long id_cliente;
 
     // Constructor
-    public Cuenta(int id_cuenta, int pin_cuenta, double saldo_cuenta, int id_cliente) {
+    public Cuenta(long id_cuenta, long pin_cuenta, BigDecimal saldo_cuenta, long id_cliente) {
         this.id_cuenta = id_cuenta;
         this.id_cliente = id_cliente;
         this.pin_cuenta = pin_cuenta;
@@ -15,35 +18,35 @@ public class Cuenta{
 
     // Getter y Setter para cada atributo
 
-    public int getId_cuenta() {
+    public long getId_cuenta() {
         return id_cuenta;
     }
 
-    public void setId_cuenta(int id_cuenta) {
+    public void setId_cuenta(long id_cuenta) {
         this.id_cuenta = id_cuenta;
     }
 
-    public int getPin_cuenta() {
+    public long getPin_cuenta() {
         return pin_cuenta;
     }
 
-    public void setPin_cuenta(int pin_cuenta) {
+    public void setPin_cuenta(long pin_cuenta) {
         this.pin_cuenta = pin_cuenta;
     }
 
-    public double getSaldo_cuenta() {
+    public BigDecimal getSaldo_cuenta() {
         return saldo_cuenta;
     }
 
-    public void setSaldo_cuenta(double saldo_cuenta) {
+    public void setSaldo_cuenta(BigDecimal saldo_cuenta) {
         this.saldo_cuenta = saldo_cuenta;
     }
 
-    public int getId_cliente() {
+    public long getId_cliente() {
         return id_cliente;
     }
 
-    public void setId_cliente(int id_cliente) {
+    public void setId_cliente(long id_cliente) {
         this.id_cliente = id_cliente;
     }
